Toss resolvers instead of rejecting state promises
The three possible states (running, success, failed) depend
on one another to some extent.
For example, if a step fails it will never succeed, and vice versa.
However, there are upsides to expose states as promises to the UI layer
so it has several promises to act upon.

A downside of that design is that we have to make sure that a step can
never fulfill conflicting states at the same time.
For example, a step should never be able to mark itself as successful
and failed, firing both promises.
To make sure a step can’t do that, we have a `preventStates` function,
which acts as an interlock by auto-rejecting conflicting promises
preemptively. That leverages the contract of Promise, which guarantees
that `rejected` is a final state for a promise.
However, that would cause unwanted side-effects as all those bogus
rejections would appear in the browser console as errors, adding
unnecessary noise that would distract from real errors.

Change the `preventStates` function to remove those unwanted effects.
Rather than preemptively auto-rejecting the conflicting promises, the
function now forgets its resolver so it can never be fulfilled.
That has a similar effect as before, plus it keeps the console clean.